+++ src/GraphicalInterface.py
@@ -1,4 +1,4 @@
-APP_VERSION = "1.0.10"
+APP_VERSION = "1.0.12"
 """
 GraphicalInterface.py - Main Application Window for ErgoProtect
 ----------------------------------------------------------------